edit & delete done

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,26 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
+
+        User loggedUser = (User) session.getAttribute("loggeduser");
+
         Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if (loggedUser.getId().intValue() == image.getUser().getId().intValue()) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            //model.addAttribute("comments", image.getComments());
+            return "images/edit";
+        }
+        else
+        {
+            model.addAttribute("image",image);
+            model.addAttribute("editError","only the owner of image can edit the image");
+           // model.addAttribute("comments",image.getComments());
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +153,22 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session,Model model) {
+        User u = (User) session.getAttribute("loggeduser");
+        Image I = imageService.getImage(imageId);
+        if (u.getId().intValue() == I.getUser().getId().intValue()) {
+
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+
+        else
+        {
+            model.addAttribute("image",I);
+            model.addAttribute("deleteError","Only the owner of the image can delete the image");
+           // model.addAttribute("comments",I.getComments());
+            return "images/image";
+        }
     }
